refactor(api): modularize routes and remove rate limiting
- index.js:
  - remove express-rate-limit middleware and related config
  - extract all REST API endpoints and error handling into separate routes module
  - simplify main file by delegating route logic via dependency injection
